news-site
This is a news-site application that collects news stories from multiple sources on various topics in real-time. It consists of a server component and a client component. It is built using the MERN-Stack.

In addition to MERN-Stack, this application leverages the NewsAPI, a popular news provider, to fetch the latest news articles from a wide range of sources. By incorporating NewsAPI, the news site offers a comprehensive and up-to-date selection of news content to keep users informed and engaged.